backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
genai
google-generativeai
MongoClient
geodesic
pymongo
geopy